Moved ApplicationInfoPopup to subfolder.
