+++ scenarioo-client/app/shared/index.ts
@@ -3,8 +3,8 @@ import './restServices';
 import './utils/urlContextExtractor.service';
 import './localStorage/localStorage.service';
 import './navigation/selectedBranchAndBuild.service';
-import './navigation/applicationInfoPopup.service';
-import './navigation/applicationInfoPopup.controller';
+import './navigation/applicationInfoPopup/applicationInfoPopup.service';
+import './navigation/applicationInfoPopup/applicationInfoPopup.controller';
 import './filterableTableTree/treeNode.service';
 import './filterableTableTree/referenceTreeNavigation.service';
 import './navigation/sharePage/sharePage.service';
