+++ scenarioo-client/app/shared/navigation/applicationInfoPopup/applicationInfoPopup.html
