+++ scenarioo-client/app/shared/navigation/applicationInfoPopup/applicationInfoPopup.service.js
